requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==1.26.4
pillow==10.3.0
keras==3.8.0
tensorflow==2.18.0